+++ envelope/internal/tests/services/cache_clearafter.go
@@ -26,6 +26,7 @@ package services
 import (
 	"fmt"
 	"go.uber.org/thriftrw/wire"
+	"go.uber.org/zap/zapcore"
 	"strings"
 )
 
@@ -146,6 +147,16 @@ func (v *Cache_ClearAfter_Args) Equals(rhs *Cache_ClearAfter_Args) bool {
 	}
 
 	return true
+}
+
+// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
+// fast logging of Cache_ClearAfter_Args.
+func (v *Cache_ClearAfter_Args) MarshalLogObject(enc zapcore.ObjectEncoder) error {
+	if v.DurationMS != nil {
+		enc.AddInt64("durationMS", *v.DurationMS)
+	}
+
+	return nil
 }
 
 // GetDurationMS returns the value of DurationMS if it is set or its