+++ gen/enum.go
@@ -125,6 +125,23 @@ func enum(g Generator, spec *compile.EnumSpec) error {
 			return []byte(<$strconv>.FormatInt(int64(<$v>), 10)), nil
 		}
 
+		<$zapcore := import "go.uber.org/zap/zapcore">
+
+		// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
+		// fast logging of <$enumName>.
+		func (<$v> <$enumName>) MarshalLogObject(enc <$zapcore>.ObjectEncoder) error {
+			enc.AddInt32("value", int32(<$v>))
+			<if len .Spec.Items ->
+				switch int32(<$v>) {
+				<range .UniqueItems ->
+					case <.Value>:
+						enc.AddString("name", "<.Name>")
+				<end ->
+				}
+			<end ->
+			return nil
+		}
+
 		// Ptr returns a pointer to this enum value.
 		func (<$v> <$enumName>) Ptr() *<$enumName> {
 			return &<$v>