+++ gen/internal/tests/services/keyvalue_getvalue.go
@@ -9,6 +9,7 @@ import (
 	"go.uber.org/thriftrw/gen/internal/tests/exceptions"
 	"go.uber.org/thriftrw/gen/internal/tests/unions"
 	"go.uber.org/thriftrw/wire"
+	"go.uber.org/zap/zapcore"
 	"strings"
 )
 
@@ -119,6 +120,16 @@ func (v *KeyValue_GetValue_Args) Equals(rhs *KeyValue_GetValue_Args) bool {
 	}
 
 	return true
+}
+
+// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
+// fast logging of KeyValue_GetValue_Args.
+func (v *KeyValue_GetValue_Args) MarshalLogObject(enc zapcore.ObjectEncoder) error {
+	if v.Key != nil {
+		enc.AddString("key", (string)(*v.Key))
+	}
+
+	return nil
 }
 
 // GetKey returns the value of Key if it is set or its
@@ -388,6 +399,23 @@ func (v *KeyValue_GetValue_Result) Equals(rhs *KeyValue_GetValue_Result) bool {
 	}
 
 	return true
+}
+
+// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
+// fast logging of KeyValue_GetValue_Result.
+func (v *KeyValue_GetValue_Result) MarshalLogObject(enc zapcore.ObjectEncoder) error {
+	if v.Success != nil {
+		if err := enc.AddObject("success", v.Success); err != nil {
+			return err
+		}
+	}
+	if v.DoesNotExist != nil {
+		if err := enc.AddObject("doesNotExist", v.DoesNotExist); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // GetSuccess returns the value of Success if it is set or its