+++ gen/set.go
@@ -20,7 +20,11 @@
 
 package gen
 
-import "go.uber.org/thriftrw/compile"
+import (
+	"fmt"
+
+	"go.uber.org/thriftrw/compile"
+)
 
 // setGenerator generates logic to convert lists of arbitrary Thrift types to
 // and from ValueLists.
@@ -196,4 +200,50 @@ func (s *setGenerator) Equals(g Generator, spec *compile.SetSpec) (string, error
 	)
 
 	return name, wrapGenerateError(spec.ThriftName(), err)
+}
+
+func (s *setGenerator) zapMarshaler(
+	g Generator,
+	spec compile.TypeSpec,
+	root *compile.SetSpec,
+	fieldValue string,
+) (string, error) {
+	name := zapperName(g, spec)
+	if err := g.EnsureDeclared(
+		`
+			type <.Name> <typeReference .Type>
+			<$zapcore := import "go.uber.org/zap/zapcore">
+			<$s := newVar "s">
+			<$v := newVar "v">
+			<$enc := newVar "enc">
+			// MarshalLogArray implements zapcore.ArrayMarshaler, enabling
+			// fast logging of <.Name>.
+			func (<$s> <.Name>) MarshalLogArray(<$enc> <$zapcore>.ArrayEncoder) error {
+				<if isHashable .Type.ValueSpec ->
+					for <$v> := range <$s> {
+				<- else ->
+					for _, <$v> := range <$s> {
+				<- end>
+					<- $encAppend := printf "%s.Append%s(%s)" $enc (zapEncoder .Type.ValueSpec) (zapMarshaler .Type.ValueSpec $v)>
+					<if (zapCanError .Type.ValueSpec) ->
+						if err := <$encAppend>; err != nil {
+							return err
+						}
+					<- else ->
+						<$encAppend>
+					<- end>
+				}
+				return nil
+			}
+			`, struct {
+			Name string
+			Type *compile.SetSpec
+		}{
+			Name: name,
+			Type: root,
+		},
+	); err != nil {
+		return "", err
+	}
+	return fmt.Sprintf("(%v)(%v)", name, fieldValue), nil
 }