+++ plugin/api/plugin_handshake.go
@@ -27,6 +27,7 @@ import (
 	"errors"
 	"fmt"
 	"go.uber.org/thriftrw/wire"
+	"go.uber.org/zap/zapcore"
 	"strings"
 )
 
@@ -141,6 +142,18 @@ func (v *Plugin_Handshake_Args) Equals(rhs *Plugin_Handshake_Args) bool {
 	}
 
 	return true
+}
+
+// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
+// fast logging of Plugin_Handshake_Args.
+func (v *Plugin_Handshake_Args) MarshalLogObject(enc zapcore.ObjectEncoder) error {
+	if v.Request != nil {
+		if err := enc.AddObject("request", v.Request); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // GetRequest returns the value of Request if it is set or its
@@ -375,6 +388,18 @@ func (v *Plugin_Handshake_Result) Equals(rhs *Plugin_Handshake_Result) bool {
 	}
 
 	return true
+}
+
+// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
+// fast logging of Plugin_Handshake_Result.
+func (v *Plugin_Handshake_Result) MarshalLogObject(enc zapcore.ObjectEncoder) error {
+	if v.Success != nil {
+		if err := enc.AddObject("success", v.Success); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // GetSuccess returns the value of Success if it is set or its