+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+- Added `zapcore.MarshalLogObject/Array` method to all generated structs 
+  and its underlying components, enums, and non-primitive typedefs, implementing 
+  Zap marshalers.
 
 ## [1.12.0] - 2018-06-25
 ### Added